+++ src/main/java/co/edu/icesi/zoo/dto/AnimalNoParentsDTO.java
@@ -18,7 +18,7 @@ public class AnimalNoParentsDTO {
     private char sex;
     private int weight;
     private int age;
-    private int height;
+    private double height;
     private String arrivalDate;
